% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemid.R
\docType{methods}
\name{clustexp}
\alias{clustexp}
\alias{clustexp,SCseq-method}
\title{Method clustexp}
\usage{
clustexp(object, clustnr = 20, bootnr = 50, metric = "pearson",
  do.gap = FALSE, sat = TRUE, SE.method = "Tibs2001SEmax",
  SE.factor = 0.25, B.gap = 50, cln = 0, rseed = 17000,
  FUNcluster = "kmedoids")

\S4method{clustexp}{SCseq}(object, clustnr = 20, bootnr = 50,
  metric = "pearson", do.gap = FALSE, sat = TRUE,
  SE.method = "Tibs2001SEmax", SE.factor = 0.25, B.gap = 50, cln = 0,
  rseed = 17000, FUNcluster = "kmedoids")
}
\description{
Method clustexp
}
