% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltree.R
\docType{methods}
\name{compentropy}
\alias{compentropy}
\alias{compentropy,Ltree-method}
\title{Method compentropy}
\usage{
compentropy(object)

\S4method{compentropy}{Ltree}(object)
}
\description{
Method compentropy
}
