% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltree.R
\docType{methods}
\name{compspantree}
\alias{compspantree}
\alias{compspantree,Ltree-method}
\title{Method compspantree}
\usage{
compspantree(object)

\S4method{compspantree}{Ltree}(object)
}
\description{
Method compspantree
}
