% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemid.R
\docType{methods}
\name{comptsne}
\alias{comptsne}
\alias{comptsne,SCseq-method}
\title{Method comptsne}
\usage{
comptsne(object, rseed = 15555, sammonmap = FALSE, initial_cmd = TRUE,
  ...)

\S4method{comptsne}{SCseq}(object, rseed = 15555, sammonmap = FALSE,
  initial_cmd = TRUE, ...)
}
\description{
Method comptsne
}
