% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemid.R
\docType{methods}
\name{filterdata}
\alias{filterdata}
\alias{filterdata,SCseq-method}
\title{Method filterdata}
\usage{
filterdata(object, mintotal = 3000, minexpr = 5, minnumber = 1,
  maxexpr = Inf, downsample = TRUE, dsn = 1, rseed = 17000)

\S4method{filterdata}{SCseq}(object, mintotal = 3000, minexpr = 5,
  minnumber = 1, maxexpr = Inf, downsample = TRUE, dsn = 1,
  rseed = 17000)
}
\description{
Method filterdata
}
