% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemid.R
\docType{methods}
\name{findoutliers}
\alias{findoutliers}
\alias{findoutliers,SCseq-method}
\title{Method findoutliers}
\usage{
findoutliers(object, outminc = 5, outlg = 2, probthr = 0.001,
  thr = 2^-(1:40), outdistquant = 0.95)

\S4method{findoutliers}{SCseq}(object, outminc = 5, outlg = 2,
  probthr = 0.001, thr = 2^-(1:40), outdistquant = 0.95)
}
\description{
Method findoutliers
}
