% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltree.R
\docType{methods}
\name{lineagetree}
\alias{lineagetree}
\alias{lineagetree,Ltree-method}
\title{Method lineagetree}
\usage{
lineagetree(object, pthr = 0.01, nmode = FALSE)

\S4method{lineagetree}{Ltree}(object, pthr = 0.01, nmode = FALSE)
}
\arguments{
\item{object}{Ltree object}

\item{pthr}{?}

\item{nmode}{?}
}
\description{
Method lineagetree
}
