% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltree.R
\docType{class}
\name{Ltree-class}
\alias{Ltree-class}
\alias{Ltree}
\alias{Ltree}
\title{Lineage tree}
\usage{
\S4method{initialize}{Ltree}(.Object, sc)
}
\description{
Lineage tree

Constructor method of class Ltree
}
\section{Slots}{

\describe{
\item{\code{sc}}{SCseq object}

\item{\code{ldata}}{?}

\item{\code{entropy}}{?}

\item{\code{trproj}}{?}

\item{\code{par}}{?}

\item{\code{prback}}{?}

\item{\code{prbacka}}{?}

\item{\code{ltcoord}}{?}

\item{\code{prtree}}{?}

\item{\code{sigcell}}{?}

\item{\code{cdata}}{?}

\item{\code{trl}}{?}
}}

