% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltree.R
\docType{methods}
\name{projback}
\alias{projback}
\alias{projback,Ltree-method}
\title{Method projback}
\usage{
projback(object, pdishuf = 2000, nmode = FALSE, rseed = 17000)

\S4method{projback}{Ltree}(object, pdishuf = 2000, nmode = FALSE,
  rseed = 17000)
}
\arguments{
\item{object}{ltree object}

\item{pdishuf}{number of pdishuffles}

\item{nmode}{if false, won't pdishuffle}

\item{rseed}{sed to be set}
}
\description{
Method projback
}
