% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltree.R
\docType{methods}
\name{projcells}
\alias{projcells}
\alias{projcells,Ltree-method}
\title{Method projcells}
\usage{
projcells(object, cthr = 0, nmode = FALSE)

\S4method{projcells}{Ltree}(object, cthr = 0, nmode = FALSE)
}
\arguments{
\item{object}{ltree object}

\item{cthr}{TODO}

\item{nmode}{TODO}
}
\description{
Method projcells
}
